+++ samples/extensions/failure-notify/send.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+echo "Residue dispatch error $1. File: $2" | mail -s 'residue err' $3