+++ src/logging/log.h
@@ -63,16 +63,34 @@
 #define RESIDUE_UNUSED(x) (void)x
 
 #ifdef RESIDUE_PROFILING
+#   define RESIDUE_PROFILE_CHECKPOINT_TEMPL(id, result, idx, refidx, type, unit) std::chrono::high_resolution_clock::time_point id##chkpnt_3##idx = std::chrono::high_resolution_clock::now();\
+       result = std::chrono::duration_cast<type>( id##chkpnt_3##idx - id##1 ).count();\
+       {auto resultforcompar = std::chrono::duration_cast<type>( id##chkpnt_3##idx - id##chkpnt_3##refidx ).count();\
+       RLOG(DEBUG) << #id << idx << " from base (" << #id << "1) " << result << " " << unit << " (from checkpoint " << #id << refidx << ": " << resultforcompar << " " << unit ")";}
 #   define RESIDUE_PROFILE_START(id) std::chrono::high_resolution_clock::time_point id##1 = std::chrono::high_resolution_clock::now(); RESIDUE_UNUSED(id##1)
 #   define RESIDUE_PROFILE_END(id, result) std::chrono::high_resolution_clock::time_point id##2 = std::chrono::high_resolution_clock::now();\
        result = std::chrono::duration_cast<std::chrono::milliseconds>( id##2 - id##1 ).count();
-#   define RESIDUE_PROFILE_CHECKPOINT(id, result, idx) std::chrono::high_resolution_clock::time_point id##3##idx = std::chrono::high_resolution_clock::now();\
-       result = std::chrono::duration_cast<std::chrono::milliseconds>( id##3##idx - id##1 ).count();\
-       std::cout << idx << " checkpoint at " << result << " ms\n";
+#   define RESIDUE_PROFILE_CHECKPOINT(id, result, idx, refidx) RESIDUE_PROFILE_CHECKPOINT_TEMPL(id, result, idx, refidx, std::chrono::milliseconds, "ms")
+#   define RESIDUE_PROFILE_CHECKPOINT_NS(id, result, idx, refidx) RESIDUE_PROFILE_CHECKPOINT_TEMPL(id, result, idx, refidx, std::chrono::nanoseconds, "nans")
+#   define RESIDUE_PROFILE_CHECKPOINT_MIS(id, result, idx, refidx) RESIDUE_PROFILE_CHECKPOINT_TEMPL(id, result, idx, refidx, std::chrono::microseconds, "mis")
+
 #else
 #   define RESIDUE_PROFILE_START(id)
 #   define RESIDUE_PROFILE_END(id, result)
-#   define RESIDUE_PROFILE_CHECKPOINT(id, result, idx)
+#   define RESIDUE_PROFILE_CHECKPOINT(id, result, idx, refidx)
+#   define RESIDUE_PROFILE_CHECKPOINT_MIS(id, result, idx, refidx)
+#   define RESIDUE_PROFILE_CHECKPOINT_NS(id, result, idx, refidx)
 #endif // RESIDUE_PROFILING
+
+#if defined(RESIDUE_DEV) && defined(RESIDUE_PROFILING)
+#define RESIDUE_HIGH_RESOLUTION_PROFILING
+#   define RESIDUE_HIGH_PROFILE_CHECKPOINT(id, result, idx, refidx) RESIDUE_PROFILE_CHECKPOINT(id, result, idx, refidx)
+#   define RESIDUE_HIGH_PROFILE_CHECKPOINT_NS(id, result, idx, refidx) RESIDUE_PROFILE_CHECKPOINT_NS(id, result, idx, refidx)
+#   define RESIDUE_HIGH_PROFILE_CHECKPOINT_MIS(id, result, idx, refidx) RESIDUE_PROFILE_CHECKPOINT_MIS(id, result, idx, refidx)
+#else
+#   define RESIDUE_HIGH_PROFILE_CHECKPOINT(id, result, idx, refidx)
+#   define RESIDUE_HIGH_PROFILE_CHECKPOINT_NS(id, result, idx, refidx)
+#   define RESIDUE_HIGH_PROFILE_CHECKPOINT_MIS(id, result, idx, refidx)
+#endif
 
 #endif // LOG_H